Split Login Form to separate client component from page
